Rename harmonics.cpp to harmonicsNode.cpp
