+++ maya/pluginMain.cpp
@@ -1,4 +1,4 @@
-#include "harmonics.h"
+#include "harmonicsNode.h"
 #include <maya/MFnPlugin.h>
 
 MStatus initializePlugin( MObject obj )
